+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/PlatformApi.java
@@ -20,7 +20,7 @@ import com.refinedmods.refinedstorage2.platform.api.network.node.exporter.Export
 import com.refinedmods.refinedstorage2.platform.api.network.node.externalstorage.PlatformExternalStorageProviderFactory;
 import com.refinedmods.refinedstorage2.platform.api.network.node.importer.ImporterTransferStrategyFactory;
 import com.refinedmods.refinedstorage2.platform.api.resource.ItemResource;
-import com.refinedmods.refinedstorage2.platform.api.resource.filter.ResourceType;
+import com.refinedmods.refinedstorage2.platform.api.resource.filter.FilteredResourceFactory;
 import com.refinedmods.refinedstorage2.platform.api.storage.StorageRepository;
 import com.refinedmods.refinedstorage2.platform.api.storage.channel.PlatformStorageChannelType;
 import com.refinedmods.refinedstorage2.platform.api.storage.type.StorageType;
@@ -61,8 +61,6 @@ public interface PlatformApi {
 
     MutableComponent createTranslation(String category, String value, Object... args);
 
-    OrderedRegistry<ResourceLocation, ResourceType> getResourceTypeRegistry();
-
     ComponentMapFactory<NetworkComponent, Network> getNetworkComponentMapFactory();
 
     OrderedRegistry<ResourceLocation, GridSynchronizer> getGridSynchronizerRegistry();
@@ -93,4 +91,8 @@ public interface PlatformApi {
                                                       GridServiceFactory gridServiceFactory);
 
     void addGridScrollingStrategyFactory(GridScrollingStrategyFactory scrollingStrategyFactory);
+
+    void addFilteredResourceFactory(FilteredResourceFactory factory);
+
+    FilteredResourceFactory getFilteredResourceFactory();
 }