+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/storage/channel/AbstractPlatformStorageChannelType.java
@@ -5,11 +5,11 @@ import com.refinedmods.refinedstorage2.api.storage.channel.StorageChannel;
 import com.refinedmods.refinedstorage2.api.storage.channel.StorageChannelType;
 import com.refinedmods.refinedstorage2.api.storage.tracked.TrackedResource;
 
-import java.util.Optional;
 import java.util.function.BiConsumer;
 
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.resources.ResourceLocation;
 import org.apiguardian.api.API;
 
 @API(status = API.Status.STABLE, since = "2.0.0-milestone.2.4")
@@ -20,13 +20,22 @@ public abstract class AbstractPlatformStorageChannelType<T> implements PlatformS
     private final String name;
     private final StorageChannelType<T> delegate;
     private final MutableComponent title;
+    private final ResourceLocation textureIdentifier;
+    private final int textureX;
+    private final int textureY;
 
     protected AbstractPlatformStorageChannelType(final String name,
                                                  final StorageChannelType<T> delegate,
-                                                 final MutableComponent title) {
+                                                 final MutableComponent title,
+                                                 final ResourceLocation textureIdentifier,
+                                                 final int textureX,
+                                                 final int textureY) {
         this.name = name;
         this.delegate = CoreValidations.validateNotNull(delegate, "Delegate cannot be null");
         this.title = title;
+        this.textureIdentifier = textureIdentifier;
+        this.textureX = textureX;
+        this.textureY = textureY;
     }
 
     @Override
@@ -37,8 +46,6 @@ public abstract class AbstractPlatformStorageChannelType<T> implements PlatformS
         return tag;
     }
 
-    protected abstract CompoundTag toTag(T resource);
-
     @Override
     public void fromTag(final CompoundTag tag, final BiConsumer<T, TrackedResource> acceptor) {
         fromTag(tag).ifPresent(resource -> {
@@ -48,8 +55,6 @@ public abstract class AbstractPlatformStorageChannelType<T> implements PlatformS
         });
     }
 
-    protected abstract Optional<T> fromTag(CompoundTag tag);
-
     @Override
     public StorageChannel<T> create() {
         return delegate.create();
@@ -58,6 +63,21 @@ public abstract class AbstractPlatformStorageChannelType<T> implements PlatformS
     @Override
     public MutableComponent getTitle() {
         return title;
+    }
+
+    @Override
+    public ResourceLocation getTextureIdentifier() {
+        return textureIdentifier;
+    }
+
+    @Override
+    public int getXTexture() {
+        return textureX;
+    }
+
+    @Override
+    public int getYTexture() {
+        return textureY;
     }
 
     @Override