+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/containermenu/ResourceTypeAccessor.java
@@ -1,9 +1,0 @@
-package com.refinedmods.refinedstorage2.platform.common.containermenu;
-
-import net.minecraft.network.chat.Component;
-
-public interface ResourceTypeAccessor {
-    Component getCurrentResourceTypeName();
-
-    void toggleResourceType();
-}