+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/containermenu/storage/AbstractStorageContainerMenu.java
@@ -1,26 +1,19 @@
 package com.refinedmods.refinedstorage2.platform.common.containermenu.storage;
 
 import com.refinedmods.refinedstorage2.api.core.filter.FilterMode;
-import com.refinedmods.refinedstorage2.api.core.registry.OrderedRegistry;
 import com.refinedmods.refinedstorage2.api.storage.AccessMode;
-import com.refinedmods.refinedstorage2.platform.api.resource.filter.ResourceType;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.AbstractResourceFilterContainerMenu;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.property.ClientProperty;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.property.PropertyTypes;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.property.ServerProperty;
-import com.refinedmods.refinedstorage2.platform.common.internal.resource.filter.ResourceFilterContainer;
 import com.refinedmods.refinedstorage2.platform.common.util.RedstoneMode;
 
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.MenuType;
 
 public abstract class AbstractStorageContainerMenu extends AbstractResourceFilterContainerMenu {
-    protected AbstractStorageContainerMenu(final MenuType<?> type,
-                                           final int syncId,
-                                           final OrderedRegistry<ResourceLocation, ResourceType> resourceTypeRegistry) {
-        super(type, syncId, resourceTypeRegistry);
-
+    protected AbstractStorageContainerMenu(final MenuType<?> type, final int syncId) {
+        super(type, syncId);
         registerProperty(new ClientProperty<>(PropertyTypes.PRIORITY, 0));
         registerProperty(new ClientProperty<>(PropertyTypes.FILTER_MODE, FilterMode.BLOCK));
         registerProperty(new ClientProperty<>(PropertyTypes.FUZZY_MODE, false));
@@ -30,12 +23,9 @@ public abstract class AbstractStorageContainerMenu extends AbstractResourceFilte
 
     protected AbstractStorageContainerMenu(final MenuType<?> type,
                                            final int syncId,
-                                           final OrderedRegistry<ResourceLocation, ResourceType> resourceTypeRegistry,
                                            final Player player,
-                                           final StorageConfigurationContainer configContainer,
-                                           final ResourceFilterContainer container) {
-        super(type, syncId, resourceTypeRegistry, player, container);
-
+                                           final StorageConfigurationContainer configContainer) {
+        super(type, syncId, player);
         registerProperty(new ServerProperty<>(
             PropertyTypes.PRIORITY,
             configContainer::getPriority,