+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/packet/ClientToServerCommunications.java
@@ -3,7 +3,6 @@ package com.refinedmods.refinedstorage2.platform.common.packet;
 import com.refinedmods.refinedstorage2.api.grid.service.GridExtractMode;
 import com.refinedmods.refinedstorage2.api.grid.service.GridInsertMode;
 import com.refinedmods.refinedstorage2.platform.api.grid.GridScrollMode;
-import com.refinedmods.refinedstorage2.platform.api.resource.filter.ResourceType;
 import com.refinedmods.refinedstorage2.platform.api.storage.channel.PlatformStorageChannelType;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.property.PropertyType;
 
@@ -24,9 +23,9 @@ public interface ClientToServerCommunications {
 
     <T> void sendPropertyChange(PropertyType<T> type, T value);
 
-    void sendResourceTypeChange(ResourceType type);
-
     void sendStorageInfoRequest(UUID storageId);
+
+    void sendResourceFilterSlotChange(int slotIndex, boolean tryAlternatives);
 
     void sendResourceFilterSlotAmountChange(int slotIndex, long amount);
 }