+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/packet/s2c/ServerToClientCommunicationsImpl.java
@@ -4,7 +4,6 @@ import com.refinedmods.refinedstorage2.api.storage.StorageInfo;
 import com.refinedmods.refinedstorage2.api.storage.tracked.TrackedResource;
 import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
 import com.refinedmods.refinedstorage2.platform.api.storage.channel.PlatformStorageChannelType;
-import com.refinedmods.refinedstorage2.platform.common.internal.resource.filter.ResourceFilterContainer;
 import com.refinedmods.refinedstorage2.platform.common.packet.ServerToClientCommunications;
 import com.refinedmods.refinedstorage2.platform.forge.packet.NetworkManager;
 
@@ -49,14 +48,31 @@ public class ServerToClientCommunicationsImpl implements ServerToClientCommunica
     }
 
     @Override
-    public void sendResourceFilterSlotUpdate(final ServerPlayer player,
-                                             final ResourceFilterContainer resourceFilterContainer,
-                                             final int slotIndex,
-                                             final int containerIndex) {
-        networkManager.send(
-            player,
-            new ResourceFilterSlotUpdatePacket(slotIndex, containerIndex, resourceFilterContainer)
-        );
+    public <T> void sendResourceFilterSlotUpdate(final ServerPlayer player,
+                                                 @Nullable final PlatformStorageChannelType<T> storageChannelType,
+                                                 @Nullable final T resource,
+                                                 final long amount,
+                                                 final int slotIndex) {
+        if (storageChannelType != null && resource != null) {
+            PlatformApi.INSTANCE
+                .getStorageChannelTypeRegistry()
+                .getId(storageChannelType)
+                .ifPresent(id -> networkManager.send(player, new ResourceFilterSlotUpdatePacket<>(
+                    slotIndex,
+                    storageChannelType,
+                    id,
+                    resource,
+                    amount
+                )));
+        } else {
+            networkManager.send(player, new ResourceFilterSlotUpdatePacket<>(
+                slotIndex,
+                null,
+                null,
+                null,
+                amount
+            ));
+        }
     }
 
     @Override